+++ docs/start/changelog.rst
@@ -11,6 +11,7 @@ ChangeLog
     * attr command will only print the contents of the replies if there's only one device
       being referenced. Makes it easier to pipe into jq
     * Add ``lifx lan:effect_off`` for turning off firmware and waveform effects on a device
+    * Removed messages that were made un-public https://github.com/LIFX/public-protocol/pull/5
 
 .. _release-core-0-40-3: